make invoke_provider_fn into a static function
