+++ compiler/rustc_middle/src/query/plumbing.rs
@@ -91,13 +91,6 @@ pub struct QueryVTable<'tcx, C: QueryCache, H: QueryHelper<'tcx, C::Key, C::Valu
     pub state: QueryState<'tcx, C::Key>,
     pub cache: C,
 
-    /// Function pointer that actually calls this query's provider.
-    /// Also performs some associated secondary tasks; see the macro-defined
-    /// implementation in `mod invoke_provider_fn` for more details.
-    ///
-    /// This should be the only code that calls the provider function.
-    pub invoke_provider_fn: fn(tcx: TyCtxt<'tcx>, key: C::Key) -> C::Value,
-
     pub helper: H,
 
     /// Function pointer that hashes this query's result values.
@@ -132,6 +125,13 @@ pub trait QueryHelper<'tcx, K, V>: Default + 'static {
     fn try_load_from_disk_fn(tcx: TyCtxt<'tcx>, prev_index: SerializedDepNodeIndex) -> Option<V>;
 
     fn will_cache_on_disk_for_key(key: K) -> bool;
+
+    /// Function pointer that actually calls this query's provider.
+    /// Also performs some associated secondary tasks; see the macro-defined
+    /// implementation in `mod invoke_provider_fn` for more details.
+    ///
+    /// This should be the only code that calls the provider function.
+    fn invoke_provider_fn(tcx: TyCtxt<'tcx>, key: K) -> V;
 }
 
 impl<'tcx, C: QueryCache, H: QueryHelper<'tcx, C::Key, C::Value>> fmt::Debug
@@ -351,8 +351,6 @@ macro_rules! define_callbacks {
                         tcx: TyCtxt<'tcx>,
                         prev_index: crate::dep_graph::SerializedDepNodeIndex,
                     ) -> Option<Erased<Value<'tcx>>> {
-                        use rustc_middle::queries::$name::{ProvidedValue, provided_to_erased};
-
                         let loaded_value: ProvidedValue<'tcx> =
                             rustc_middle::query::plumbing::try_load_from_disk(tcx, prev_index)?;
 
@@ -382,6 +380,47 @@ macro_rules! define_callbacks {
                         }
                     }
 
+                    fn invoke_provider_fn(tcx: TyCtxt<'tcx>, key: Key<'tcx>) -> Erased<Value<'tcx>> {
+                        invoke_provider_fn::__rust_begin_short_backtrace(tcx, key)
+                    }
+                }
+
+                /// Defines an `invoke_provider` function that calls the query's provider,
+                /// to be used as a function pointer in the query's vtable.
+                ///
+                /// To mark a short-backtrace boundary, the function's actual name
+                /// (after demangling) must be `__rust_begin_short_backtrace`.
+                mod invoke_provider_fn {
+                    use super::*;
+
+                    #[inline(never)]
+                    pub(crate) fn __rust_begin_short_backtrace<'tcx>(
+                        tcx: TyCtxt<'tcx>,
+                        key: Key<'tcx>,
+                    ) -> Erased<Value<'tcx>> {
+                        #[cfg(debug_assertions)]
+                        let _guard = tracing::span!(tracing::Level::TRACE, stringify!($name), ?key).entered();
+
+                        // Call the actual provider function for this query.
+
+                        #[cfg($separate_provide_extern)]
+                        let provided_value = if let Some(local_key) = crate::query::AsLocalQueryKey::as_local_key(&key) {
+                            (tcx.query_system.local_providers.$name)(tcx, local_key)
+                        } else {
+                            (tcx.query_system.extern_providers.$name)(tcx, key)
+                        };
+
+                        #[cfg(not($separate_provide_extern))]
+                        let provided_value = (tcx.query_system.local_providers.$name)(tcx, key);
+
+                        rustc_middle::ty::print::with_reduced_queries!({
+                            tracing::trace!(?provided_value);
+                        });
+
+                        // Erase the returned value, because `QueryVTable` uses erased values.
+                        // For queries with `arena_cache`, this also arena-allocates the value.
+                        provided_to_erased(tcx, provided_value)
+                    }
                 }
 
                 pub type Cache<'tcx> =
